feat: basic player movement
This diff wires up the action tracker such that it drives the player
sprite position.

Helps with #3 but does not close it.

Still need to fix the movement to the tile grid and check to see if the next
tile is walkable or not. Also need to set the initial position of the
player using the data in the ldtk level.