feat: movement and collision via LDtK enum data
Fixes #3

**TL;DR**: A "TileType" enum in LDtK can now be assigned to tiles in the
tileset, then on the Rust side we can identify the type of tile by grid
coordinate after pulling the enum associations for each tile out of the
JSON representation of the level.

In the `player_movement` system we can then lookup what type of tile is
currently occupied (to adjust the movement speed) then check the type of
tile the move _would place the player in_ so we can decide if the move
should be permitted.

---

During this work, some _problems_ were identified (to be resolved in
follow-up PRs).

LDtK allows you to "stack tiles" which is to say you can set multiple
tile graphics on a single grid coordinate. This breaks some assumptions
in this initial implementation since the TileType mapping is based on
array offsets in the JSON data. When tiles are "stacked" the number of
tiles in the JSON data is larger than the number of grid coordinates,
and also it becomes possible for a single coordinate to have multiple
(possibly conflicting) TileTypes.

The "stacked tiles" problem therefore requires a change in how the
TileType data is retrieved from JSON and how the storage/lookups are
done in Rust. It also could suggest that concepts like "walkable" or
"wall" should not be registered directly on the graphic tile, but rather
as separate entity data. Either way, the current implementation is
insufficient when tiles are stacked.